split off index generation into separate file
